ANDROID: KVM: arm64: Fix owner id type in IOMMU callbacks
pKVM calls into the IOMMU driver when changing ownership of a page, so
make sure to use the right type for the owner_id parameter in those
callbacks.

XXX: we should expose the owner_id type instead of hardcoding u32.

+++ arch/arm64/kvm/hyp/nvhe/iommu/s2mpu.c
@@ -237,7 +237,7 @@ static void set_mpt_range_locked(struct mpt *mpt, phys_addr_t first_byte,
 		__range_invalidation(dev, first_byte, last_byte);
 }
 
-static void s2mpu_host_stage2_set_owner(phys_addr_t addr, size_t size, u8 owner_id)
+static void s2mpu_host_stage2_set_owner(phys_addr_t addr, size_t size, u32 owner_id)
 {
 	/* Grant access only to the default owner of the page table (ID=0). */
 	enum mpt_prot prot = owner_id ? MPT_PROT_NONE : MPT_PROT_RW;
